(PC-23083)[PRO] refactor: clean ico-warning-grey
