--- pro/src/components/InfoDialog/index.ts
@@ -1,1 +1,0 @@
-export { default } from './InfoDialog'
