+++ pro/src/components/VenueForm/ReimbursementPoint/ReimbursementPoint.module.scss
@@ -73,3 +73,7 @@
     width: rem.torem(10px);
   }
 }
+
+.dialog-confirm-button {
+  margin-top: rem.torem(16px);
+}
